# Conflicts:
#	accounts/app_interface/service_interface.py
#	accounts/graphql/queries.py
#	accounts/graphql/types/types.py
#	accounts/interactors/address/address_interactor.py
#	accounts/interactors/populate_data/import_addresses.py
#	accounts/interactors/populate_data/import_users.py
#	accounts/interactors/storage_interface/address_storage_interface.py
#	accounts/models/user.py
#	accounts/storages/address_storage.py
#	accounts/storages/user_storage.py
#	accounts/tests/factories/storage_factories.py
#	orders/exception/custom_exceptions.py
#	orders/graphql/types/error_types.py
#	orders/interactors/order/order_interactor.py
#	orders/interactors/order/place_order_interactor.py
#	orders/interactors/order/update_order_interactor.py
#	orders/interactors/populate_data/import_promo_codes.py
#	orders/mixin/order_mixin.py
#	orders/models/order.py
#	orders/storages/order_storage.py
#	orders/tests/factories/storage_factories.py
#	restaurants/adapter/order.py
#	restaurants/graphql/mutations/cart/clear_cart_items_mutation.py
#	restaurants/graphql/mutations/cart/remove_cart_item_mutation.py
#	restaurants/graphql/resolvers/get_cart_items_resolver.py
#	restaurants/graphql/types/input_types.py
#	restaurants/graphql/types/types.py
#	restaurants/interactors/populate_data/import_menu_items.py
#	restaurants/interactors/populate_data/import_restaurants.py
#	restaurants/interactors/restaurant/browse_restaurants.py
#	restaurants/mixins/restaurant_timing_mixin.py
#	restaurants/models/cart.py
#	restaurants/models/restaurant.py
#	restaurants/models/restaurant_timing.py
#	restaurants/storages/cart_storage.py
#	restaurants/storages/restaurant_storage.py
#	restaurants/storages/restaurant_timing_storage.py
#	utils/read_csv_util.py
#	zomato/middleware.py
#	zomato/settings.py
#	zomato/urls.py

+++ orders/graphql/types/error_types.py
@@ -6,8 +6,7 @@ class PromoCodeUsageLimitReached(graphene.ObjectType):
 
 
 class PromoCodeNotEligible(graphene.ObjectType):
-    # TODO: min_order_value is Decimal in the DTO but exposed here as Float — lossy and inconsistent with items_total below.
-    min_order_value = graphene.Float(required=True)
+    min_order_value = graphene.Decimal(required=True)
     items_total = graphene.Decimal(required=True)
 
 
@@ -39,7 +38,6 @@ class RestaurantNotOpenAtScheduledTime(graphene.ObjectType):
 
 
 class PromoCodeNotFound(graphene.ObjectType):
-    # TODO: every other error type has required=True — inconsistent contract.
     promo_code_id = graphene.Int()
